fix(test): move static images from nested directory
