+++ src/avatar/__tests__/adorable.png
