+++ src/avatar/__tests__/avatar.scenario.tsx
@@ -7,7 +7,7 @@ LICENSE file in the root directory of this source tree.
 import * as React from 'react';
 
 import { Avatar } from '..';
-import imageFile from './static/adorable.png';
+import imageFile from './adorable.png';
 
 export function Scenario() {
   return (
