+++ src/skeleton/__tests__/adorable.png
